rename alchemy engines directory
